requirements.txt
langchain_community 
langchain_text_splitters 
chromadb
pypdf
google-genai